+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## [1.1.0] - Upcoming
 
+### Added
+- New event attribute `VSLast3Avg` (Sampled VS) for stabilized vertical speed tracking
+- New issue type `AppHighVsAvgBelow1000AGL` for high sampled descent rate (<-1150 fpm) below 1000 AGL
+
+### Changed
+- Updated `AppHighVsBelow1000AGL` threshold from -1000 fpm to -1500 fpm
+
 ## [1.0.0] - 2026-01-31
 
 - Initial version